proper(?) handling of density matrices in a loop
trying to take care of more than one density matrix in a loop.
Actually, it shouldn't happen usually, because it destroys the
proper singlet spin summation and relies on
\gamma^t\alpha_u\alpha = \gamma^t\beta_u\beta = \half \gamma^t_u
But in the case of transforming one-electron h matrix to fock matrix
we have to introduce such terms.
Hopefully it won't shoot in the foot one day...